eclipse changes + split to eh1 instead of ehx1
