+++ README.md
@@ -116,8 +116,10 @@ WD-Firmware
     - From 'Eclipse MCU' menu bar select '*Project'* -> *'Build All'*. Note that you can select which platform to build for.
     - Since the building process use SCons build system, you can build via console/terminal. Please read the readme on ***’/build’*** 
 
-### Downloading & debugging the firmware image (FTDI over USB)
-- #### Setting up the hardware (taken from SiFive Freedom Studio Manual v1p6).
+### Platforms Downloading & debugging 
+- We provide several platforms to work with, please follow the instructions for the one you preferred.
+
+    #### Setting up Hifive1 - FTDI over USB (taken from SiFive Freedom Studio Manual v1p6).
 	- Connect to your HiFive1 debug interface and type "lsusb" to see if FT2232C is connected:
 
             $ lsusb
@@ -134,8 +136,7 @@ WD-Firmware
             $ groups
             ... plugdev ...
 	- Power off/on Debian station
- - #### Setting up Nexys-A7 for SweRV
-    Note: If you are not using SweRV core with Nexys-A7 you can skip this section.
+    #### Setting up Nexys-A7 for SweRV
     
     Since Nexys-A7 is an FPGA platform it need special handling...
     - ***Prerequisite***: Following are prerequisite running SweRV core on Xilinx FPGA on Nexys-A7 board
@@ -150,29 +151,28 @@ WD-Firmware
                 G4 = TMS
                 G2 = nRST
 
-    - ***FPGA file loading***: for loading the FPGA bit file, do the following steps:
+    - ***FPGA image file loading***: for loading the FPGA bit file, do the following steps:
     	- Copy the FPGA bit file /WD-Firmware/board/nexys_a7_swerv1/***swerv_eh1_reference_design.bit***
 	   to uSD device (locate it at the uSD root)
     	- Attach the uSD device to the Nexys-A7 board (uSD slot is on board's bottom)
 		- Set the following jumpers:  JP1 - connect JTAG & USB/SD pins.   JP2 - connect the 2 pins on 'SD' side
-		- That's it! From now on, at any power-on the FPGA bit file is loaded to the FPGA.
+		- At power-on the FPGA bit file is loaded to the FPGA. LED 'Busy' should be ORANGE while flushing is done.
 
-
-
+    #### Setting up ISS (works as simulator for EH1)
+    
 - #### Eclipse MCU configuration:
     - From the 'Eclipse MCU' menu bar press File->Properties->C/C++ Build->Settings, select the *Toolchain path* with bin folder: [Toolchain-root]/bin
 
     
 - #### Eclipse MCU Debug:
     - Select from the ***'Eclipse MCU'*** menu bar ***'Run' -> 'Debug Configurations...'***; 
-    - Choose the platform you wish to runs on, from **'GDB OpenOCD Debugging'** menu
-    - Current support
-
-            - hifive1
-            - nexys_A7_Swerv1
-            
-
-
+    - Choose the platform you wish to runs on, from **'left main windows'** menu
+    - Current support:
+        ```javascript
+        - hifive1                              <-- HiFive Eval board
+        - nexys_a7_Swerv1_eh1                  <-- Nexys A7 digilent FPGA board running SweRV EH1
+        - whisper_eh1_connect_and_debug        <-- SweRV ISS simulator 
+        ```
 ### Adding new source modules
 
 The folder WD-Firmware/demo/build/ contains a template file (SConscript_template) which can be used.
