+++ WD-Firmware/board/nexys_a7_swerv_eh1/swerv_openocd.cfg
@@ -1,0 +1,47 @@
+echo  "\n"
+echo  "--------------------------------------------------------------------"
+echo  "openocd cfg running: Connecting and downloading image to flash "
+echo  "--------------------------------------------------------------------"
+echo  "\n"
+
+# JTAG adapter setup
+adapter_khz     10000
+#1500
+
+
+interface ftdi
+ftdi_device_desc "Olimex OpenOCD JTAG ARM-USB-TINY-H"
+#ftdi_device_desc "Dual RS232-HS"
+ftdi_vid_pid 0x15ba 0x002a
+#ftdi_vid_pid 0x0403 0x6010
+
+echo  "\n"
+echo  "-----------------------------------------------------"
+echo  "Using FDTI device Olimex OpenOCD JTAG ARM-USB-TINY-H"
+echo  "Vendor ID = 0x15ba"
+echo  "Product ID = 0x002a"
+echo  "-----------------------------------------------------"
+echo  "\n"
+ftdi_layout_init 0x0808 0x0a1b
+ftdi_layout_signal nSRST -oe 0x0200
+ftdi_layout_signal LED -data 0x0800
+
+transport select jtag
+
+set _CHIPNAME riscv
+
+jtag newtap $_CHIPNAME cpu -irlen 5 -expected-id 0x01
+
+
+set _TARGETNAME $_CHIPNAME.cpu
+target create $_TARGETNAME riscv -chain-position $_TARGETNAME
+
+riscv set_prefer_sba on
+
+init
+#halt
+#flash protect 0 64 last off
+echo "-------------------------------------------------------------------------"
+echo "GDB Start running PLEASE hold few sec, for -load- command to complete..."
+echo "-------------------------------------------------------------------------"
+
