--- WD-Firmware/board/whisper/whisper_ehx_debug.launch
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<launchConfiguration type="org.eclipse.debug.core.groups.GroupLaunchConfigurationType">
-<stringAttribute key="org.eclipse.debug.core.launchGroup.0.action" value="NONE"/>
-<booleanAttribute key="org.eclipse.debug.core.launchGroup.0.adoptIfRunning" value="false"/>
-<booleanAttribute key="org.eclipse.debug.core.launchGroup.0.enabled" value="true"/>
-<stringAttribute key="org.eclipse.debug.core.launchGroup.0.mode" value="run"/>
-<stringAttribute key="org.eclipse.debug.core.launchGroup.0.name" value="whisper_run"/>
-<stringAttribute key="org.eclipse.debug.core.launchGroup.1.action" value="NONE"/>
-<booleanAttribute key="org.eclipse.debug.core.launchGroup.1.adoptIfRunning" value="false"/>
-<booleanAttribute key="org.eclipse.debug.core.launchGroup.1.enabled" value="true"/>
-<stringAttribute key="org.eclipse.debug.core.launchGroup.1.mode" value="inherit"/>
-<stringAttribute key="org.eclipse.debug.core.launchGroup.1.name" value="whisper_ehx"/>
-<listAttribute key="org.eclipse.debug.ui.favoriteGroups">
-<listEntry value="org.eclipse.debug.ui.launchGroup.debug"/>
-<listEntry value="org.eclipse.debug.ui.launchGroup.run"/>
-</listAttribute>
-</launchConfiguration>
