--- WD-Firmware/board/whisper/whisper_run.launch
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<launchConfiguration type="org.eclipse.ui.externaltools.ProgramLaunchConfigurationType">
-<stringAttribute key="org.eclipse.ui.externaltools.ATTR_LOCATION" value="${workspace_loc:/WD-Firmware/board/whisper/whisper}"/>
-<stringAttribute key="org.eclipse.ui.externaltools.ATTR_TOOL_ARGUMENTS" value="--target=../../demo/build/output/nexys_a7_swerv_ehx1_freertos.elf --gdb --gdb-tcp-port=7000 --startpc=0x0 --configfile=${workspace_loc:/WD-Firmware/board/whisper/whisper.json}"/>
-<stringAttribute key="org.eclipse.ui.externaltools.ATTR_WORKING_DIRECTORY" value="${workspace_loc:/WD-Firmware/board/whisper}"/>
-</launchConfiguration>
