+++ WD-Firmware/psp/api_inc/psp_swerv_eh1_csrs.h
@@ -1,0 +1,63 @@
+/* 
+* SPDX-License-Identifier: Apache-2.0
+* Copyright 2019 Western Digital Corporation or its affiliates.
+* 
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+* 
+* http:*www.apache.org/licenses/LICENSE-2.0
+* 
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+/**
+* @file   psp_swerv_eh1_csrs.h
+* @author Nati Rapaport
+* @date   01.12.2019
+* @brief  Definitions of Swerv's (EH1 version) CSRs
+* 
+*/
+#ifndef  __PSP_SWERV_EH1_CSRS_H__
+#define  __PSP_SWERV_EH1_CSRS_H__
+
+/**
+* include files
+*/
+
+/**
+* definitions
+*/
+#define M_PSP_DISABLE_SWERV_TIMER()
+#define M_PSP_ENABLE_SWERV_TIMER()
+
+/**
+* macros
+*/
+
+/**
+* types
+*/
+
+/**
+* local prototypes
+*/
+
+/**
+* external prototypes
+*/
+
+/**
+* global variables
+*/
+
+/**
+* APIs
+*/
+
+
+
+#endif /* __PSP_SWERV_EH1_CSRS_H__ */
