--- WD-Firmware/psp/psp_timer_swerv_ehx1.c
@@ -1,40 +1,0 @@
-/*
-* SPDX-License-Identifier: Apache-2.0
-* Copyright 2019 Western Digital Corporation or its affiliates.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-* http:*www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
-/**
-* @file   psp_timer_swerv_ehx1.c
-* @author Nati Rapaport
-* @date   8.12.2019
-* @brief  This file implements SweRV_EHX1 timer service functions
-*
-*/
-
-/*-----------------------------------------------------------
- * Implementation of functions defined in portable.h for the RISC-V RV32 port.
- *----------------------------------------------------------*/
-
-/**
-*
-* @brief Setup function for Timer0 (per SweRV-EHX1 reference manual)
-*
-***************************************************************************************************/
-void pspTimerSwervEhx1SetupSingleRun(const unsigned int enableInterrupt)
-{
-}
-
-
-
-
