--- WD-Firmware/rtos/rtosal/config/nexys_a7_swerv_ehx1/FreeRTOSConfig.h
@@ -1,189 +1,0 @@
-/*
-    FreeRTOS V8.2.3 - Copyright (C) 2015 Real Time Engineers Ltd.
-    All rights reserved
-
-    VISIT http://www.FreeRTOS.org TO ENSURE YOU ARE USING THE LATEST VERSION.
-
-    This file is part of the FreeRTOS distribution.
-
-    FreeRTOS is free software; you can redistribute it and/or modify it under
-    the terms of the GNU General Public License (version 2) as published by the
-    Free Software Foundation >>>> AND MODIFIED BY <<<< the FreeRTOS exception.
-
-    ***************************************************************************
-    >>!   NOTE: The modification to the GPL is included to allow you to     !<<
-    >>!   distribute a combined work that includes FreeRTOS without being   !<<
-    >>!   obliged to provide the source code for proprietary components     !<<
-    >>!   outside of the FreeRTOS kernel.                                   !<<
-    ***************************************************************************
-
-    FreeRTOS is distributed in the hope that it will be useful, but WITHOUT ANY
-    WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-    FOR A PARTICULAR PURPOSE.  Full license text is available on the following
-    link: http://www.freertos.org/a00114.html
-
-    ***************************************************************************
-     *                                                                       *
-     *    FreeRTOS provides completely free yet professionally developed,    *
-     *    robust, strictly quality controlled, supported, and cross          *
-     *    platform software that is more than just the market leader, it     *
-     *    is the industry's de facto standard.                               *
-     *                                                                       *
-     *    Help yourself get started quickly while simultaneously helping     *
-     *    to support the FreeRTOS project by purchasing a FreeRTOS           *
-     *    tutorial book, reference manual, or both:                          *
-     *    http://www.FreeRTOS.org/Documentation                              *
-     *                                                                       *
-    ***************************************************************************
-
-    http://www.FreeRTOS.org/FAQHelp.html - Having a problem?  Start by reading
-    the FAQ page "My application does not run, what could be wrong?".  Have you
-    defined configASSERT()?
-
-    http://www.FreeRTOS.org/support - In return for receiving this top quality
-    embedded software for free we request you assist our global community by
-    participating in the support forum.
-
-    http://www.FreeRTOS.org/training - Investing in training allows your team to
-    be as productive as possible as early as possible.  Now you can receive
-    FreeRTOS training directly from Richard Barry, CEO of Real Time Engineers
-    Ltd, and the world's leading authority on the world's leading RTOS.
-
-    http://www.FreeRTOS.org/plus - A selection of FreeRTOS ecosystem products,
-    including FreeRTOS+Trace - an indispensable productivity tool, a DOS
-    compatible FAT file system, and our tiny thread aware UDP/IP stack.
-
-    http://www.FreeRTOS.org/labs - Where new FreeRTOS products go to incubate.
-    Come and try FreeRTOS+TCP, our new open source TCP/IP stack for FreeRTOS.
-
-    http://www.OpenRTOS.com - Real Time Engineers ltd. license FreeRTOS to High
-    Integrity Systems ltd. to sell under the OpenRTOS brand.  Low cost OpenRTOS
-    licenses offer ticketed support, indemnification and commercial middleware.
-
-    http://www.SafeRTOS.com - High Integrity Systems also provide a safety
-    engineered and independently SIL3 certified version for use in safety and
-    mission critical applications that require provable dependability.
-
-    1 tab == 4 spaces!
-*/
-
-
-#ifndef FREERTOS_CONFIG_H
-#define FREERTOS_CONFIG_H
-
-/*#include "platform.h"*/
-
-/*
- * For some reason the standard demo timer demo/test tasks fail when executing
- * in QEMU, although they pass on other RISC-V platforms.  This requires
- * further investigation, but for now, defining _WINDOWS_ has the effect of
- * using the wider timer test thresholds that are normally only used when the
- * tests are used with the FreeRTOS Windows port (which is not deterministic
- * and therefore requires wider margins).
- */
-#define _WINDOWS_
-
-/*-----------------------------------------------------------
- * Application specific definitions.
- *
- * These definitions should be adjusted for your particular hardware and
- * application requirements.
- *
- * THESE PARAMETERS ARE DESCRIBED WITHIN THE 'CONFIGURATION' SECTION OF THE
- * FreeRTOS API DOCUMENTATION AVAILABLE ON THE FreeRTOS.org WEB SITE.
- *
- * See http://www.freertos.org/a00110.html.
- *----------------------------------------------------------*/
-
-#define configCLINT_BASE_ADDRESS		        CLINT_CTRL_ADDR
-#define configUSE_PREEMPTION			        1
-#define configCPU_CLOCK_HZ				        265000000
-#ifdef D_CLOCK_RATE
-   #define configRTC_CLOCK_HZ D_CLOCK_RATE      /* If the clock rate is already defined, take it here */
-#else
-   #define configRTC_CLOCK_HZ 32768
-#endif
-#ifdef D_TICK_TIME_MS                           /* If D_TICK_TIME_MS is already defined, use it to calculate configTICK_RATE_HZ */
-   #define configTICK_RATE_HZ				        1000/D_TICK_TIME_MS
-#else
-   #define configTICK_RATE_HZ				        250
-#endif
-#define configMAX_PRIORITIES			        3
-#define configMINIMAL_STACK_SIZE		        450 /* [NR] To-do check why in new config it is set to 2*70 */
-#define configMAX_TASK_NAME_LEN			        16
-#define configUSE_16_BIT_TICKS			        0
-#define configIDLE_SHOULD_YIELD			        1
-#define configUSE_TASK_NOTIFICATIONS            1
-#define configUSE_MUTEXES				        1
-#define configQUEUE_REGISTRY_SIZE		        10
-#define configUSE_QUEUE_SETS                    0
-#define configUSE_TIME_SLICING                  1
-#define configUSE_NEWLIB_REENTRANT              0
-#define configENABLE_BACKWARD_COMPATIBILITY     0
-#define configUSE_RECURSIVE_MUTEXES		        0
-#define configUSE_APPLICATION_TASK_TAG	        0
-#define configUSE_COUNTING_SEMAPHORES	        1
-#define configUSE_PORT_OPTIMISED_TASK_SELECTION 1
-#define configNUM_THREAD_LOCAL_STORAGE_POINTERS 5
-
-/* Memory allocation related definitions. */
-#ifndef configSUPPORT_STATIC_ALLOCATION
-   #define configSUPPORT_STATIC_ALLOCATION      1
-#endif
-
-#ifndef configSUPPORT_DYNAMIC_ALLOCATION
-   #define configSUPPORT_DYNAMIC_ALLOCATION     0
-#endif
-
-#define configTOTAL_HEAP_SIZE                   11*1024
-#define configAPPLICATION_ALLOCATED_HEAP        0
-
-/* Hook function related definitions. */
-#define configUSE_IDLE_HOOK                     1
-#define configUSE_TICK_HOOK                     1
-#define configCHECK_FOR_STACK_OVERFLOW          1
-#define configUSE_MALLOC_FAILED_HOOK            1
-#define configUSE_DAEMON_TASK_STARTUP_HOOK      0
-
-/* Run time and task stats gathering related definitions. */
-#define configGENERATE_RUN_TIME_STATS           0
-#define configUSE_TRACE_FACILITY                0
-#define configUSE_STATS_FORMATTING_FUNCTIONS    0
-
-/* Co-routine definitions. */
-#define configUSE_CO_ROUTINES 			0
-#define configMAX_CO_ROUTINE_PRIORITIES 1
-
-/* Software timer definitions. */
-#define configUSE_TIMERS				1
-#define configTIMER_TASK_PRIORITY		3
-#define configTIMER_QUEUE_LENGTH		5
-#define configTIMER_TASK_STACK_DEPTH	configMINIMAL_STACK_SIZE
-
-/* Set the following definitions to 1 to include the API function, or zero
-to exclude the API function. */
-#define INCLUDE_vTaskPrioritySet                1
-#define INCLUDE_uxTaskPriorityGet               1
-#define INCLUDE_vTaskDelete                     1
-#define INCLUDE_vTaskSuspend                    1
-#define INCLUDE_xResumeFromISR                  1
-#define INCLUDE_vTaskDelayUntil                 1
-#define INCLUDE_vTaskDelay                      1
-#define INCLUDE_xTaskGetSchedulerState          1
-#define INCLUDE_xTaskGetCurrentTaskHandle       1
-#define INCLUDE_uxTaskGetStackHighWaterMark     1
-#define INCLUDE_xTaskGetIdleTaskHandle          1
-#define INCLUDE_eTaskGetState                   1
-#define INCLUDE_xEventGroupSetBitFromISR        1
-#define INCLUDE_xTimerPendFunctionCall          1
-#define INCLUDE_xTaskAbortDelay                 1
-#define INCLUDE_xTaskGetHandle                  1
-#define INCLUDE_xTaskResumeFromISR              1
-#define INCLUDE_vTaskCleanUpResources	        1
-
-
-/* Define to trap errors during development. */
-#define configASSERT( x ) if( ( x ) == 0 ) {taskDISABLE_INTERRUPTS(); for( ;; );}
-
-
-#endif /* FREERTOS_CONFIG_H */
